third_party/WebKit/LayoutTests/media/video-controls-captions-load-by-lang.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
    <title>Test non-default track load according to user language preference.</title>
    <script src=media-file.js></script>
    <script src=media-controls.js></script>
    <script src=video-test.js></script>
    <script>
        var track;

        function startTest()
        {
            if (!window.eventSender) {
                consoleWrite("No eventSender found.");
                failTest();
            }

            findMediaElement();
            testClosedCaptionsButtonVisibility(true);

            consoleWrite("");
            consoleWrite("** The captions track should be listed in textTracks, but disabled. **");
            testExpected("video.textTracks.length", 2);
            testExpected("video.textTracks[0].mode", "disabled");
            testExpected("video.textTracks[1].mode", "disabled");

            consoleWrite("");
            consoleWrite("** Set the user language preference so that the track will be chosen when the CC button is clicked. **");
            run("internals.setUserPreferredLanguages(['ar'])");
            clickCCButton();
        }

        function loaded()
        {
            findMediaElement();
            waitForEvent('canplaythrough', startTest);

            video.src = findMediaFile('video', 'content/counting');
        }
    </script>
</head>
<body onload="loaded()">
    <p>Tests that appropriate language track is loaded, according to user preferences.</p>
    <video controls>
        <track src="track/captions-webvtt/captions-fast.vtt" kind="captions">
        <track src="track/captions-webvtt/captions-fast.vtt" kind="captions" srclang="ar" onload="endTest()">
    </video>
</body>
</html>
