third_party/WebKit/LayoutTests/media/video-controls-captions-load-by-lang-expected.txt

Tests that appropriate language track is loaded, according to user preferences.

EVENT(canplaythrough)

** Caption button should be visible and enabled.
EXPECTED (captionsButtonCoordinates[0] > '0') OK
EXPECTED (captionsButtonCoordinates[1] > '0') OK
EXPECTED (captionsButtonElement.disabled == 'false') OK

** The captions track should be listed in textTracks, but disabled. **
EXPECTED (video.textTracks.length == '2') OK
EXPECTED (video.textTracks[0].mode == 'disabled') OK
EXPECTED (video.textTracks[1].mode == 'disabled') OK

** Set the user language preference so that the track will be chosen when the CC button is clicked. **
RUN(internals.setUserPreferredLanguages(['ar']))
*** Click the CC button.
END OF TEST
